cookbook/web_scrape/minify_inscriptis_text/requirements.txt
httpx
inscriptis
minify_html
selectolax
# ogbujipt
toolio
google-re2
tiktoken